Moving to multichecker model so that each check can be its own analysis tool
