+++ analysis/timeafter/time_after.go
@@ -12,29 +12,38 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package analysis
+package timeafter
 
 import (
 	"fmt"
 	"go/ast"
-	"go/token"
+	"strings"
 
 	"golang.org/x/tools/go/analysis"
 	"golang.org/x/tools/go/analysis/passes/inspect"
 	"golang.org/x/tools/go/ast/inspector"
 )
 
-const Doc = `This is a custom vet check for the cilium codebase.`
+const (
+	// Doc for the timeafter check
+	Doc = `This is checks for "time.After" instances in for loops.`
 
+	timeAfterPkg  = "time"
+	timeAfterFunc = "After"
+)
+
+// Analyzer is the global for the multichecker
 var Analyzer = &analysis.Analyzer{
-	Name:     "cilium",
+	Name:     "timeafter",
 	Doc:      Doc,
 	Requires: []*analysis.Analyzer{inspect.Analyzer},
 	Run:      run,
 }
 
-type checker interface {
-	check(n ast.Node) (bool, string, token.Pos)
+var ignoreArg string
+
+func init() {
+	Analyzer.Flags.StringVar(&ignoreArg, "ignore", "", `list of packages to ignore ("inctimer,time")`)
 }
 
 type visitor func(ast.Node) bool
@@ -52,20 +61,51 @@ func run(pass *analysis.Pass) (interface{}, error) {
 		return nil, fmt.Errorf("analyzer is not type *inspector.Inspector")
 	}
 
-	nodeFilter := []ast.Node{
-		(*ast.ForStmt)(nil),
-		(*ast.File)(nil),
-		(*ast.ImportSpec)(nil),
+	ignoreMap := make(map[string]struct{})
+	for _, ign := range strings.Split(ignoreArg, ",") {
+		ignoreMap[strings.TrimSpace(ign)] = struct{}{}
 	}
-	checks := []checker{
-		&timeAfterCheck{},
-	}
+
+	var (
+		aliases    []string
+		ignore     = false
+		nodeFilter = []ast.Node{
+			(*ast.ForStmt)(nil),
+			(*ast.File)(nil),
+			(*ast.ImportSpec)(nil),
+		}
+	)
 	inspct.Preorder(nodeFilter, func(n ast.Node) {
-		for _, c := range checks {
-			ok, report, pos := c.check(n)
-			if !ok {
-				pass.Reportf(pos, report)
+		switch stmt := n.(type) {
+		case *ast.File:
+			_, ignore = ignoreMap[stmt.Name.Name]
+			aliases = []string{timeAfterPkg}
+		case *ast.ImportSpec:
+			if ignore {
+				return
 			}
+			// Collect aliases.
+			pkg := stmt.Path.Value
+			if pkg == fmt.Sprintf("%q", timeAfterPkg) {
+				if stmt.Name != nil {
+					aliases = append(aliases, stmt.Name.Name)
+				}
+			}
+		case *ast.ForStmt:
+			if ignore {
+				return
+			}
+			ast.Walk(visitor(func(node ast.Node) bool {
+				switch expr := node.(type) {
+				case *ast.CallExpr:
+					for _, pkg := range aliases {
+						if isPkgDot(expr.Fun, pkg, timeAfterFunc) {
+							pass.Reportf(node.Pos(), "use of %s.After in a for loop is prohibited, use inctimer instead", pkg)
+						}
+					}
+				}
+				return true
+			}), stmt.Body)
 		}
 	})
 	return nil, nil
