+++ analysis/timeafter/time_after_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package analysis
+package timeafter
 
 import (
 	"testing"
