+++ go.mod
@@ -2,4 +2,4 @@ module github.com/cilium/customvet
 
 go 1.15
 
-require golang.org/x/tools v0.0.0-20201207204333-a835c872fcea
+require golang.org/x/tools v0.0.0-20201208233053-a543418bbed2
